refactor(ts): extract timeline/normalizers.ts from store
Move createEmptyTrack, createDefaultTracks, normalizeTrack(s) and the pure
clamp helpers (clampPercent, clampTier9, clampEquipmentRefineTier) into a
dependency-free timeline/normalizers.ts module, mirroring the optimizer's
structure. No behavior change.
